move buildCsv from utils to make-csv
